Extract turn processing into session_turn, leaving SessionManager as orchestration + facade
